test : 학습율 계산 테스트 작성

+++ src/main/java/toyproject/studyscheduler/domain/study/StudyTime.java
@@ -36,6 +36,6 @@ public class StudyTime extends BaseEntity {
     }
 
     public double calculateLearningRate() {
-        return Math.round(((totalCompleteTime /study.getTotalExpectedTime())*100)*100)/100.0;
+        return Math.round((((float) totalCompleteTime /study.getTotalExpectedTime())*100)*100)/100.0;
     }
 }
